+++ app/components/screen/Widget.js
@@ -27,7 +27,9 @@ import {
   RadioPrizes,
   StackedSlider,
   StackedRadio,
+  StabilityTracker,
 } from '../../widgets';
+import ABTrails from '../../widgets/ABTrails';
 import TimePicker from '../../widgets/TimeRange/TimePicker';
 import { setSelected } from '../../state/responses/responses.actions';
 import { currentAppletSelector } from '../../state/app/app.selectors';
@@ -37,7 +39,20 @@ import {
 
 // const TOKEN_LOGGER_SCHEMA = 'https://raw.githubusercontent.com/ChildMindInstitute/TokenLogger_applet/master/protocols/TokenLogger/TokenLogger_schema';
 
-const Widget = ({ screen, answer, onChange, applet, isCurrent, isSelected, setSelected, onPress, onRelease, onContentError, appletTokenBalance }) => {
+const Widget = ({
+  screen,
+  answer,
+  onChange,
+  applet,
+  isCurrent,
+  currentScreen,
+  isSelected,
+  setSelected,
+  onPress,
+  onRelease,
+  onContentError,
+  appletTokenBalance
+}) => {
   const valueType = R.path(['valueConstraints', 'valueType'], screen);
 
   if (screen.inputType === 'radio'
@@ -68,6 +83,16 @@ const Widget = ({ screen, answer, onChange, applet, isCurrent, isSelected, setSe
     );
   }
 
+  if (screen.inputType == 'stabilityTracker') {
+    return (
+      <StabilityTracker
+        onChange={onChange}
+        config={screen.inputs}
+        isCurrent={isCurrent}
+      />
+    )
+  }
+
   if (screen.inputType === 'stackedRadio') {
     return (
       <StackedRadio
@@ -91,6 +116,16 @@ const Widget = ({ screen, answer, onChange, applet, isCurrent, isSelected, setSe
         value={answer}
       />
     );
+  }
+  if (screen.inputType === 'trail') {
+    return (
+      <ABTrails
+        screen={screen.variableName}
+        currentScreen={currentScreen}
+        data={answer}
+        onChange={onChange}
+      />
+    )
   }
   if (screen.inputType === 'stackedSlider') {
     return (
@@ -316,6 +351,7 @@ Widget.propTypes = {
   setSelected: PropTypes.func.isRequired,
   isCurrent: PropTypes.bool.isRequired,
   isSelected: PropTypes.bool.isRequired,
+  currentScreen: PropTypes.number.isRequired,
   onPress: PropTypes.func,
   onRelease: PropTypes.func,
   appletTokenBalance: PropTypes.object.isRequired,