+++ app/services/activityNavigation.js
@@ -1,5 +1,6 @@
 import i18n from 'i18next';
 import Screen from '../components/screen';
+import { screens } from '../widgets/ABTrails/TrailsData';
 
 const NEXT = i18n.t('activity_navigation:next');
 const SKIP = i18n.t('activity_navigation:skip');
@@ -8,7 +9,19 @@ const BACK = i18n.t('activity_navigation:back');
 const RETURN = i18n.t('activity_navigation:return');
 const UNDO = i18n.t('activity_navigation:undo');
 
-export const checkValidity = (item, response) => Screen.isValid(response, item);
+export const checkValidity = (item, response, index, tutorialStatus = 0) => { 
+  if (item.inputType === "trail" && tutorialStatus !== 0) {
+    return true;
+  }
+  if (item.inputType === "trail" && index >= 0 && response) {
+    const screen = screens[item.inputType + '' + (index + 1)];
+
+    if (screen.items.length !== response.value.currentIndex) {
+      return false;
+    }
+  }
+  return Screen.isValid(response, item);
+}
 
 export const checkSkippable = (activity, item) => {
   if (activity.skippable === true) {
@@ -48,9 +61,20 @@ export const getLastPos = (index, ar) => {
   return -1;
 };
 
-export const getNextLabel = (index, isSplashScreen, visibility, activity, responses, isContentError) => {
+export const getNextLabel = (
+  index,
+  isSplashScreen,
+  visibility,
+  activity,
+  responses,
+  isContentError,
+  tutorialStatus
+) => {
   // If the screen is not valid, then the label is Skip
-  const isValid = checkValidity(activity.items[index], responses[index]);
+  const isValid = checkValidity(activity.items[index], responses[index], index, tutorialStatus);
+  if (activity.items[index].inputType === "trail" && tutorialStatus === 1) {
+    return i18n.t('activity_navigation:skip');
+  }
 
   if (isSplashScreen) {
     return i18n.t('activity_navigation:next');
@@ -70,13 +94,16 @@ export const getNextLabel = (index, isSplashScreen, visibility, activity, respon
 };
 
 // If item has a valid response, or is skippable, then next is enabled
-export const isNextEnabled = (index, activity, responses) => {
-  const isValid = checkValidity(activity.items[index], responses[index]);
+export const isNextEnabled = (index, activity, responses, tutorialStatus) => {
+  const isValid = checkValidity(activity.items[index], responses[index], index, tutorialStatus);
   const isSkippable = checkSkippable(activity, activity.items[index]);
   return isValid || isSkippable;
 };
 
 export const isPrevEnabled = (index, activity) => {
+  if (activity.items[index].inputType === "trail") {
+    return false;
+  }
   if (activity.items[index].backDisabled === true) {
     return false;
   }