+++ app/state/app/app.selectors.js
@@ -14,6 +14,8 @@ export const connectionSelector = R.path(['app', 'isConnected']);
 
 export const appStatusSelector = R.path(['app', 'appStatus']);
 
+export const tutorialStatusSelector = R.path(['app', 'isTutorial']);
+
 export const lastActiveTimeSelector = R.path(['app', 'lastActive']);
 
 export const activities = R.path(['app', 'activities']);