+++ app/state/responses/responses.selectors.js
@@ -113,3 +113,16 @@ export const itemVisiblitySelector = createSelector(
     });
   }
 );
+
+export const challengePhaseLambdaSelector = createSelector(
+  currentResponsesSelector,
+  ({ responses, screenIndex }) => {
+    for (let i = 0; i < screenIndex; i++) {
+      if (responses[i] && responses[i].phaseType == 'challenge-phase') {
+        return responses[i].maxLambda;
+      }
+    }
+
+    return 0;
+  }
+);