+++ app/widgets/ABTrails/TrailsData.js
@@ -1,0 +1,318 @@
+
+export const screens = {
+  trail1: {
+    r: 14,
+    fontSize: 20,
+    colors: {
+      passed: 'green',
+      pending: 'black',
+      failed: 'red',
+      label: 'black',
+    },
+    items: [
+      {
+        order: 1,
+        cx: 175,
+        cy: 40,
+        label: '1',
+      },
+      {
+        order: 2,
+        cx: 285,
+        cy: 60,
+        label: '2',
+      },
+      {
+        order: 3,
+        cx: 295,
+        cy: 250,
+        label: '3',
+      },
+      {
+        order: 4,
+        cx: 180,
+        cy: 290,
+        label: '4',
+      },
+      {
+        order: 5,
+        cx: 260,
+        cy: 135,
+        label: '5',
+      },
+      {
+        order: 6,
+        cx: 57,
+        cy: 130,
+        label: '6',
+      },
+      {
+        order: 7,
+        cx: 180,
+        cy: 190,
+        label: '7',
+      },
+      {
+        order: 8,
+        cx: 65,
+        cy: 265,
+        label: '8',
+      },
+      {
+        order: 9,
+        cx: 50,
+        cy: 200,
+        label: '9',
+      },
+      {
+        order: 10,
+        cx: 170,
+        cy: 105,
+        label: '10',
+      },
+      {
+        order: 11,
+        cx: 45,
+        cy: 50,
+        label: '11',
+      },
+    ]
+  },
+
+  trail2: {
+    r: 14,
+    fontSize: 20,
+    colors: {
+      passed: 'green',
+      pending: 'black',
+      failed: 'red',
+      label: 'black',
+    },
+    items: [
+      {
+        order: 1,
+        cx: 57,
+        cy: 130,
+        label: '1',
+      },
+      {
+        order: 2,
+        cx: 260,
+        cy: 135,
+        label: '2',
+      },
+      {
+        order: 3,
+        cx: 175,
+        cy: 40,
+        label: '3',
+      },
+      {
+        order: 4,
+        cx: 45,
+        cy: 50,
+        label: '4',
+      },
+      {
+        order: 5,
+        cx: 180,
+        cy: 190,
+        label: '5',
+      },
+      {
+        order: 6,
+        cx: 285,
+        cy: 60,
+        label: '6',
+      },
+      {
+        order: 7,
+        cx: 295,
+        cy: 250,
+        label: '7',
+      },
+      {
+        order: 8,
+        cx: 50,
+        cy: 200,
+        label: '8',
+      },
+      {
+        order: 9,
+        cx: 65,
+        cy: 265,
+        label: '9',
+      },
+      {
+        order: 10,
+        cx: 170,
+        cy: 105,
+        label: '10',
+      },
+      {
+        order: 11,
+        cx: 180,
+        cy: 290,
+        label: '11',
+      },
+    ]
+  },
+
+  trail3: {
+    r: 14,
+    fontSize: 20,
+    colors: {
+      passed: 'green',
+      pending: 'black',
+      failed: 'red',
+      label: 'black',
+    },
+    items: [
+      {
+        order: 1,
+        cx: 180,
+        cy: 290,
+        label: '1',
+      },
+      {
+        order: 2,
+        cx: 295,
+        cy: 250,
+        label: 'A',
+      },
+      {
+        order: 3,
+        cx: 285,
+        cy: 60,
+        label: '2',
+      },
+      {
+        order: 4,
+        cx: 175,
+        cy: 40,
+        label: 'B',
+      },
+      {
+        order: 5,
+        cx: 260,
+        cy: 135,
+        label: '3',
+      },
+      {
+        order: 6,
+        cx: 50,
+        cy: 200,
+        label: 'C',
+      },
+      {
+        order: 7,
+        cx: 170,
+        cy: 105,
+        label: '4',
+      },
+      {
+        order: 8,
+        cx: 45,
+        cy: 50,
+        label: 'D',
+      },
+      {
+        order: 9,
+        cx: 180,
+        cy: 190,
+        label: '5',
+      },
+      {
+        order: 10,
+        cx: 57,
+        cy: 130,
+        label: 'E',
+      },
+      {
+        order: 11,
+        cx: 65,
+        cy: 265,
+        label: '6',
+      },
+    ]
+  },
+
+  trail4: {
+    r: 14,
+    fontSize: 20,
+    colors: {
+      passed: 'green',
+      pending: 'black',
+      failed: 'red',
+      label: 'black',
+    },
+    items: [
+      {
+        order: 1,
+        cx: 170,
+        cy: 105,
+        label: '1',
+      },
+      {
+        order: 2,
+        cx: 45,
+        cy: 50,
+        label: 'A',
+      },
+      {
+        order: 3,
+        cx: 175,
+        cy: 40,
+        label: '2',
+      },
+      {
+        order: 4,
+        cx: 57,
+        cy: 130,
+        label: 'B',
+      },
+      {
+        order: 5,
+        cx: 260,
+        cy: 135,
+        label: '3',
+      },
+      {
+        order: 6,
+        cx: 285,
+        cy: 60,
+        label: 'C',
+      },
+      {
+        order: 7,
+        cx: 295,
+        cy: 250,
+        label: '4',
+      },
+      {
+        order: 8,
+        cx: 50,
+        cy: 200,
+        label: 'D',
+      },
+      {
+        order: 9,
+        cx: 180,
+        cy: 190,
+        label: '5',
+      },
+      {
+        order: 10,
+        cx: 65,
+        cy: 265,
+        label: 'E',
+      },
+      {
+        order: 11,
+        cx: 180,
+        cy: 290,
+        label: '6',
+      },
+    ]
+  },
+}