+++ app/widgets/index.js
@@ -1,5 +1,7 @@
+export * from './ABTrails';
 export * from './AudioRecord/AudioImageRecord';
 export * from './AudioRecord';
+export * from './AgeSelector';
 export * from './AudioStimulus';
 export * from './Camera';
 export * from './DatePicker';
@@ -20,3 +22,4 @@ export * from './TooltipBox';
 export * from './StackedSlider';
 export * from './AgeSelector';
 export * from './StackedRadio';
+export * from './StabilityTracker';